requirements.txt
PyMuPDF
numpy
requests
spire.presentation
python-pptx
pywin32
pywinauto
scikit-image
Pillow
windnd
python-dotenv
